SIM: Move BagManager to python package
