+++ simulation/sub8_gazebo/models/sub8/sub8.sdf
@@ -3,12 +3,12 @@
   <model name="sub8">
     <link name="base_link">
       <velocity_decay>
-        <linear>0.06</linear>
+        <linear>0.01</linear>
         <angular>0.06</angular>
       </velocity_decay>
 
       <inertial>
-        <mass>26.29</mass>
+        <mass>35.29</mass>
         <inertia>
           <ixx>1.18</ixx>
           <ixy>-0.003</ixy>
@@ -178,7 +178,7 @@
 
     <plugin name="sub8_buoyancy" filename="libsub8_buoyancy.so">
       <fluid_density>1000</fluid_density>
-      <drag_coefficient>5</drag_coefficient>
+      <drag_coefficient>1</drag_coefficient>
       <link name="base_link">
         <center_of_volume>0 0 0</center_of_volume>
         <volume>0.037</volume>
