+++ simulation/sub8_gazebo/sub8_gazebo_tools/__init__.py
@@ -1,0 +1,1 @@
+from bag_manager import BagManager
